Vash Initial Public Commit
Due to the chaotic nature of Vash's development, having spent time as
part of a larger project, as a personal side project, and a few places
in between, it is infeasible for us to disentangle the development
history of Vash from everything it has touched. Because of this we have
decided to start this repository off from a fresh commit, free from the
entanglements of historical circumstance.