compute total for guided random walk on the fly
We used to pre-compute these, for performance.  This
is not going to be possible with each channel of the
tree using a different subset.  We may also need, at
some point, to modify the node frequencies on the fly.
We can revisit this later if we find a performance
bottleneck here.